# docs/swagger.yaml
definitions:
  pagamento.Pagamento:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  pagamento.PaymentCallback:
    properties:
      action:
        type: string
      api_version:
        type: string
      data:
        properties:
          id:
            type: string
        type: object
      date_created:
        type: string
      id:
        type: integer
      live_mode:
        type: boolean
      type:
        type: string
      user_id:
        type: integer
    type: object
  pagamento.QRCode:
    properties:
      qr_code:
        type: string
    type: object
info:
  contact:
    email: support@fastfood.io
    name: API Support
    url: http://www.fastfood.io/support
  description: Here you will find everything you need to have the best possible integration
    with our APIs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://fastfood.io/terms/
  title: Payments API
  version: "1.0"
paths:
  /pagamentos/{id}:
    get:
      operationId: get-payment-by-id
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagamento.Pagamento'
        "404":
          description: Not Found
      summary: Get payment by ID
      tags:
      - Payments
  /pagamentos/{id}/qr-code:
    get:
      operationId: get-qr-code-by-id
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagamento.QRCode'
        "404":
          description: Not Found
      summary: Get QR Code pagamento
      tags:
      - Payments
  /pagamentos/health:
    get:
      operationId: health-check
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Health check
      tags:
      - Payments
  /pagamentos/mp-webhook:
    post:
      operationId: receive-callback
      parameters:
      - description: Order data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pagamento.PaymentCallback'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagamento.Pagamento'
        "400":
          description: Bad Request
      summary: Receive payment callback from MercadoPago
      tags:
      - Payments
swagger: "2.0"